Give up CTest.
Now all tests are performed by one executable command, and I would like to see it's output when I run 'make test' to find out which tests are failed.
I haven't found how to make CTest to print the output, so I replaced it with custom 'test' target which runs the GTest binary.