Make config_t non-copyable.
Since config_t doesn't hold the underlying dynamic_t object, it's unsafe to make copies of it.
The example. An application parses own config with config_parser_t and does some setup. Then it needs to configure its logging library. The application calls special function from the library and provides a config_t object to it. But the library is buggy and copies the config_t object in order to parse it later when the underlying dynamic_t object is already dead. The application crashes, admins are saying that Elliptics doesn't work.
Now this is disallowed. The logging library cannot copy config_t. If it's going to delay configuration parsing, it has no choice but copy the underlying dynamic_t object.